Refactor download with less layers
